Addresses 1792 refs[t:1792] Further implementation of show engine status
git-svn-id: file:///svn/toku/tokudb@14742 c7de825b-a66e-492c-adef-691d508d4ae1

+++ newbrt/checkpoint.h
@@ -67,5 +67,5 @@ int toku_checkpoint(CACHETABLE ct, TOKULOGGER logger, char **error_string,
  * (If checkpoint is in progress, it may overwrite status info while it is being read.)
  *****/
 
-u_int64_t toku_checkpoint_get_footprint(void);
+u_int32_t toku_checkpoint_get_footprint(void);
